requirements.txt
requests==2.32.3
google-cloud-pubsub==2.29.0